refactor(ui): reorganización de la estructura de carpetas de componentes
- Se renombraron algunos archivos para mayor claridad y consistencia.
- No se realizaron cambios en la lógica ni en lo visual.
