+++ Front-End/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './pages/home/home.component';
-import { ExperienceHistoryComponent } from './pages/experience-history/experience-history.component';
+import { HomeComponent } from './components/home/home-section/home.component';
+import { ExperienceHistoryComponent } from './components/experience-history/experience-history-section/experience-history.component';
 
 const routes: Routes = [
   {path:'', component: HomeComponent},
