+++ Front-End/src/app/components/experience-history/card-experience/card-experience.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Experience } from '../../models/experience.model';
+import { Experience } from 'src/app/models/experience.model';
 
 @Component({
   selector: 'app-card-experience',
