+++ Front-End/src/app/components/experience-history/carousel-experiences/carousel-experiences.component.css
