+++ Front-End/src/app/components/home/card-actividad/card-actividad.component.css
